requirements.txt
discord.py
torch
cupy-cuda11x
tensorflow
